chore: 域名切换 + sync_engine 拆分 + 分类图标 user-global + MCP/TestFlight 提示
- 官网与帮助文档域名:youths.cc → count.beejz.com(issue 模板 / 源码引用)
- lib/cloud/sync/sync_engine.dart 按职责拆成 7 个 part 文件:apply /
  attachments / profile / realtime / resolvers / serialization / status
- 分类自定义图标:走 user-global 上传 endpoint,server 端不再按账本复制
  一份;同步页拆出"分类图标"独立计数指标
- ProductPromo 弹窗:iOS + 有 TestFlight URL + 已上架时,把 [TestFlight 内测]
  和 [前往应用商店] 并排放进 content 顶部,[前往官网] 单独居中放底部
- ProductPromo 数据结构加 testFlightUrl 字段;beeAssets 填上 TestFlight 链接
- README 顶部加 MCP support 一行式提示(简化版,只留链接)
- l10n 加 productPromoTestFlight 键(zh/zh-TW/en)
