+++ .github/ISSUE_TEMPLATE/01-bug_report_zh.yml
@@ -15,7 +15,7 @@ body:
     attributes:
       label: 前置检查
       options:
-        - label: 我已查阅 [官方文档](https://beecount-website.pages.dev/docs/intro),确认不是使用方式问题
+        - label: 我已查阅 [官方文档](https://count.beejz.com/docs/intro),确认不是使用方式问题
           required: true
         - label: 我已搜过 [已有 Issues](https://github.com/TNT-Likely/BeeCount/issues?q=is%3Aissue),没有找到重复
           required: true
