+++ .github/ISSUE_TEMPLATE/02-feature_request_zh.yml
@@ -13,7 +13,7 @@ body:
     attributes:
       label: 前置检查
       options:
-        - label: 我已查阅 [官方文档](https://beecount-website.pages.dev/docs/intro),确认该功能尚未实现
+        - label: 我已查阅 [官方文档](https://count.beejz.com/docs/intro),确认该功能尚未实现
           required: true
         - label: 我已搜过 [已有 Issues / Discussions](https://github.com/TNT-Likely/BeeCount/issues?q=is%3Aissue),没有重复
           required: true
