+++ .github/ISSUE_TEMPLATE/03-question_zh.yml
@@ -15,7 +15,7 @@ body:
     attributes:
       label: 前置检查
       options:
-        - label: 我已查阅 [官方文档](https://beecount-website.pages.dev/docs/intro),没有找到答案
+        - label: 我已查阅 [官方文档](https://count.beejz.com/docs/intro),没有找到答案
           required: true
         - label: 我已搜过 [Issues](https://github.com/TNT-Likely/BeeCount/issues) 和 [Discussions](https://github.com/TNT-Likely/BeeCount/discussions),没有找到相同问题
           required: true
