+++ lib/l10n/app_en.arb
@@ -1760,6 +1760,7 @@
   "productPromoAndroidTitle": "Request Beta Access",
   "productPromoAndroidMessage": "This app is still in closed testing on Google Play — invitation only.\n\nHow to apply: email us with your Google account address (required) and a brief use case (optional). We'll reply within 1-3 days and add you to the beta whitelist.",
   "productPromoOpenStore": "Open in App Store",
+  "productPromoTestFlight": "TestFlight beta",
   "productPromoLearnMore": "Pro",
   "productPromoEmailLabel": "Application Email (tap to copy)",
   "productPromoCopiedToast": "Email copied to clipboard",
@@ -2728,6 +2729,7 @@
   "syncHealthGroupAll": "All ledgers",
   "syncHealthRowTx": "Transactions",
   "syncHealthRowAttachment": "Attachments",
+  "syncHealthRowCategoryIcon": "Category icons",
   "syncHealthRowBudget": "Budgets",
   "syncHealthRowAccount": "Accounts",
   "syncHealthRowCategory": "Categories",
