+++ lib/l10n/app_zh.arb
@@ -1654,6 +1654,7 @@
   "productPromoAndroidTitle": "申请加入内测",
   "productPromoAndroidMessage": "这款 App 还在 Google Play 内测阶段,需要邀请才能下载。\n\n申请方式:发邮件给我们,告诉我们你的 Google 账号邮箱(必填),以及简单说明使用场景(可选)。我们会在 1-3 天内回复并加你到内测白名单。",
   "productPromoOpenStore": "前往应用商店",
+  "productPromoTestFlight": "TestFlight 内测",
   "productPromoLearnMore": "Pro",
   "productPromoEmailLabel": "申请邮箱(点击复制)",
   "productPromoCopiedToast": "邮箱已复制到剪贴板",
@@ -2688,6 +2689,7 @@
   "syncHealthGroupAll": "全部账本",
   "syncHealthRowTx": "交易",
   "syncHealthRowAttachment": "附件",
+  "syncHealthRowCategoryIcon": "分类图标",
   "syncHealthRowBudget": "预算",
   "syncHealthRowAccount": "账户",
   "syncHealthRowCategory": "分类",
