+++ lib/l10n/app_zh_TW.arb
@@ -10,6 +10,7 @@
   "productPromoAndroidTitle": "申請加入內測",
   "productPromoAndroidMessage": "這款 App 還在 Google Play 內測階段,需要邀請才能下載。\n\n申請方式:寄信給我們,告訴我們你的 Google 帳號信箱(必填),以及簡單說明使用場景(選填)。我們會在 1-3 天內回覆並加你到內測白名單。",
   "productPromoOpenStore": "前往應用商店",
+  "productPromoTestFlight": "TestFlight 內測",
   "productPromoLearnMore": "Pro",
   "productPromoEmailLabel": "申請信箱(點擊複製)",
   "productPromoCopiedToast": "信箱已複製到剪貼簿",
