+++ lib/pages/cloud/beecount_cloud_sync_page.dart
@@ -314,6 +314,7 @@ class _BeeCountCloudSyncPageState extends ConsumerState<BeeCountCloudSyncPage> {
           totalTx: SyncCountPair.missing(),
           totalAttachments: SyncCountPair.missing(),
           totalBudgets: SyncCountPair.missing(),
+          categoryAttachments: SyncCountPair.missing(),
           accounts: SyncCountPair.missing(),
           categories: SyncCountPair.missing(),
           tags: SyncCountPair.missing(),
@@ -362,6 +363,7 @@ class _BeeCountCloudSyncPageState extends ConsumerState<BeeCountCloudSyncPage> {
         _groupHeader(context, l10n.syncHealthGroupAll),
         _pairRow(context, l10n.syncHealthRowTx, effective.totalTx),
         _pairRow(context, l10n.syncHealthRowAttachment, effective.totalAttachments),
+        _pairRow(context, l10n.syncHealthRowCategoryIcon, effective.categoryAttachments),
         _pairRow(context, l10n.syncHealthRowBudget, effective.totalBudgets),
         _pairRow(context, l10n.syncHealthRowAccount, effective.accounts),
         _pairRow(context, l10n.syncHealthRowCategory, effective.categories),
